消息包头flag字段由16bit改为8bit

+++ Billow/Eddy/Net/PackageHeadFlags.cs
@@ -6,7 +6,7 @@ using System.Text;
 namespace Eddy.Net
 {
     [Flags]
-    public enum PackageHeadFlags : ushort
+    public enum PackageHeadFlags : byte
     {
         Partial = 0x0001,     // 包不完全标志，即实际包长 > ushort.MaxValue需要拆包发送
     }
